feat: make installation more reliable using computed checksums

+++ crates/tracel-mlir-rs/src/diagnostic/severity.rs
@@ -14,10 +14,10 @@ pub enum DiagnosticSeverity {
 }
 
 // Cast bindgen constants to u32 so this works on targets where they're i32.
-const SEV_ERROR: u32 = MlirDiagnosticSeverity_MlirDiagnosticError as u32;
-const SEV_NOTE: u32 = MlirDiagnosticSeverity_MlirDiagnosticNote as u32;
-const SEV_REMARK: u32 = MlirDiagnosticSeverity_MlirDiagnosticRemark as u32;
-const SEV_WARN: u32 = MlirDiagnosticSeverity_MlirDiagnosticWarning as u32;
+const SEV_ERROR: u32 = MlirDiagnosticSeverity_MlirDiagnosticError;
+const SEV_NOTE: u32 = MlirDiagnosticSeverity_MlirDiagnosticNote;
+const SEV_REMARK: u32 = MlirDiagnosticSeverity_MlirDiagnosticRemark;
+const SEV_WARN: u32 = MlirDiagnosticSeverity_MlirDiagnosticWarning;
 
 #[inline]
 fn from_raw_u32(severity: u32) -> Result<DiagnosticSeverity, Error> {
